finish depth map generation

+++ examples/gen-l-video/glv/models/unet.py
@@ -356,7 +356,6 @@ class UNet3DConditionModel(nn.Cell):
         # timesteps does not contain any weights and will always return f32 tensors
         # but time_embedding might actually be running in fp16. so we need to cast here.
         # there might be better ways to encapsulate this.
-        t_emb = t_emb.to(dtype=self.dtype)
         emb = self.time_embedding(t_emb)
 
         if self.class_embedding is not None:
@@ -366,7 +365,7 @@ class UNet3DConditionModel(nn.Cell):
             if self.class_embed_type == "timestep":
                 class_labels = self.time_proj(class_labels)
 
-            class_emb = self.class_embedding(class_labels).to(dtype=self.dtype)
+            class_emb = self.class_embedding(class_labels)
             emb = emb + class_emb
 
         # pre-process
